refactor(hosted-media): move shared types out of mod
